rccl: move AINIC recovery into separate file
