+++ projects/rccl/src/transport/net_ib_cast/CMakeLists.txt
@@ -8,6 +8,8 @@ set(NET_IB_CAST_SOURCES
     src/transport/net_ib_cast/p2p_cast.h
     src/transport/net_ib_cast/p2p_resiliency_cast.h
     src/transport/net_ib_cast/p2p_resiliency_recovery_cast.h
+    src/transport/net_ib_cast/p2p_resiliency_recovery_internal.h
+    src/transport/net_ib_cast/p2p_resiliency_recovery_ainic.h
     src/transport/net_ib_cast/common.cc
     src/transport/net_ib_cast/connect.cc
     src/transport/net_ib_cast/gdr.cc
@@ -15,6 +17,7 @@ set(NET_IB_CAST_SOURCES
     src/transport/net_ib_cast/p2p.cc
     src/transport/net_ib_cast/p2p_resiliency.cc
     src/transport/net_ib_cast/p2p_resiliency_recovery.cc
+    src/transport/net_ib_cast/p2p_resiliency_recovery_ainic.cc
     src/transport/net_ib_cast/reg.cc
     src/transport/net_ib_cast/scheduler.cc
     # gin.cc and gin_host_gdaki.cc require gin/gin_host.h (CUDA/NCCL-only) — exclude from ROCm build.
